uploads/requirements.txt
# Core runtime dependencies
PySide6
pyautogui
pyperclip
opencv-python
numpy
PyYAML
pynput
Pillow

# Optional: text OCR support used by locate_text_pytesseract
pytesseract